MVP-18: added class LoadFileProtocol
